requirements.txt
dice~=2.4.2
discord.py~=1.3.4
youtube-dl>=2020.11.26